rename packages/create DTO package for new class name convention
