+++ src/main/java/com/samrice/readingroomapi/domains/Book.java
@@ -1,4 +1,4 @@
-package com.samrice.readingroomapi.domain;
+package com.samrice.readingroomapi.domains;
 
 import org.springframework.data.annotation.Id;
