+++ src/main/java/com/samrice/readingroomapi/dtos/SearchedAuthorDto.java
@@ -1,0 +1,6 @@
+package com.samrice.readingroomapi.dtos;
+
+import java.util.List;
+
+public record SearchedAuthorDto(String key, String name, String birthDate, String deathDate, String topWork, Integer workCount, List<String> topSubjects) {
+}
