+++ src/main/java/com/samrice/readingroomapi/dtos/SearchedAuthorResponseDto.java
@@ -1,0 +1,4 @@
+package com.samrice.readingroomapi.dtos;
+
+public record SearchedAuthorResponseDto() {
+}
