--- src/main/java/com/samrice/readingroomapi/librarysearchresponses/SearchedAuthor.java
@@ -1,6 +1,0 @@
-package com.samrice.readingroomapi.librarysearchresponses;
-
-import java.util.List;
-
-public record SearchedAuthor(String key, String name, String birthDate, String deathDate, String topWork, Integer workCount, List<String> topSubjects) {
-}
