+++ src/main/java/com/samrice/readingroomapi/repositories/UserRepository.java
@@ -1,6 +1,6 @@
 package com.samrice.readingroomapi.repositories;
 
-import com.samrice.readingroomapi.domain.User;
+import com.samrice.readingroomapi.domains.User;
 import com.samrice.readingroomapi.exceptions.RrAuthException;
 
 public interface UserRepository {
