+++ src/main/java/com/samrice/readingroomapi/services/BookServiceImpl.java
@@ -3,7 +3,7 @@ package com.samrice.readingroomapi.services;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.samrice.readingroomapi.domain.Book;
+import com.samrice.readingroomapi.domains.Book;
 import com.samrice.readingroomapi.exceptions.RrBadRequestException;
 import com.samrice.readingroomapi.exceptions.RrResourceNotFoundException;
 import com.samrice.readingroomapi.repositories.BookRepository;
