+++ src/main/java/com/samrice/readingroomapi/services/LibrarySearchServiceImpl.java
@@ -1,22 +1,18 @@
 package com.samrice.readingroomapi.services;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.samrice.readingroomapi.Constants;
 import com.samrice.readingroomapi.exceptions.RrBadRequestException;
-import com.samrice.readingroomapi.librarysearchresponses.SearchedAuthor;
+import com.samrice.readingroomapi.dtos.SearchedAuthorDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -26,25 +22,24 @@ public class LibrarySearchServiceImpl implements LibrarySearchService {
     RestTemplate restTemplate;
 
     @Override
-    public List<SearchedAuthor> searchAuthors(String authorName) throws RrBadRequestException {
+    public List<SearchedAuthorDto> searchAuthors(String authorName, Integer pageNo, Integer pageSize) throws RrBadRequestException {
         try {
             ObjectMapper mapper = new ObjectMapper();
             String endpoint = Constants.OPEN_LIBRARY_SEARCH_BASE_URL + "/authors.json?q=" + authorName;
             ResponseEntity<String> response = restTemplate.getForEntity(endpoint, String.class);
             JsonNode searchResponseRoot = mapper.readTree(response.getBody());
             List<LinkedHashMap> authorsResponse = mapper.convertValue(searchResponseRoot.get("docs"), List.class);
-            List<SearchedAuthor> authorsList = authorsResponse
+            List<SearchedAuthorDto> authorsList = authorsResponse
                 .stream()
                 .filter(a -> Integer.parseInt(a.get("work_count").toString()) != 0)
                 .map(a -> mapToSearchedAuthor(a)).toList();
             return authorsList;
         } catch (Exception e) {
-            e.printStackTrace();
             throw new RrBadRequestException(e.getMessage());
         }
     }
 
-    private SearchedAuthor mapToSearchedAuthor(LinkedHashMap author) {
+    private SearchedAuthorDto mapToSearchedAuthor(LinkedHashMap author) {
         String key = author.get("key").toString();
         String name = author.get("name").toString();
         String birthDate = author.get("birth_date") != null ? author.get("birth_date").toString() : null;
@@ -52,6 +47,6 @@ public class LibrarySearchServiceImpl implements LibrarySearchService {
         String topWork = author.get("top_work") != null ? author.get("top_work").toString() : null;
         Integer workCount = Integer.parseInt(author.get("work_count").toString());
         List<String> topSubjects = (List<String>) author.get("top_subjects");
-        return new SearchedAuthor(key, name, birthDate, deathDate, topWork, workCount, topSubjects);
+        return new SearchedAuthorDto(key, name, birthDate, deathDate, topWork, workCount, topSubjects);
     }
 }
