+++ src/test/java/com/samrice/readingroomapi/repositories/ShelfRepositoryIntegrationTests.java
@@ -1,7 +1,7 @@
 package com.samrice.readingroomapi.repositories;
 
-import com.samrice.readingroomapi.domain.Book;
-import com.samrice.readingroomapi.domain.Shelf;
+import com.samrice.readingroomapi.domains.Book;
+import com.samrice.readingroomapi.domains.Shelf;
 import com.samrice.readingroomapi.exceptions.RrBadRequestException;
 import com.samrice.readingroomapi.exceptions.RrResourceNotFoundException;
 import org.junit.jupiter.api.Test;
